requirements.txt
langchain==0.1.4
langchain_community==0.0.16
openai==0.27.10
pymongo==4.5.0
python-dotenv==1.0.1
streamlit==1.28.0
